Bug#13406172 - HAVING CLAUSE ON AGGREGATED COLUMN RETURNS                WRONG RESULT WITH MYISAM
For queries with aggregate functions but no group by clause, 
a single row with aggregate functions using NULL should be 
returned even if no rows matched the conditions. However,
the having clause still has to be evaluated an may filter out
this single row.

The bug was that MySQL did not evaluate the having clause for
queries with exactly 1 row in a MyISAM table (system table) and
with a WHERE clause that is considered expensive 
(is_expensive() is true). Examples: stored procedures and 
materialized subqueries. This is a corner-case issue: 
* If the where clause had not been expensive, 
  make_join_select() would have evaluated it and noticed that 
  it is false which would lead to return_zero_rows() being 
  called. return_zero_rows() checks the having clause 
  correctly. 
* If there had been a row satisfy WHERE, do_select() would 
  have called end_send_group() which also checks the having 
  clause.