BUG#11762697 (MySQL55323): ADD AN OPTION TO LIMIT THE WRITE RATE TO SLOW QUERY LOG
DBA may now set a maximum number of "Query does not use index"
warnings per minute that may appear in the slow query log.
Limit may be set at start (--log-throttle-queries-not-using-indexes=...)
and can be read/modified at runtime using a global variable
(SET GLOBAL log_throttle_queries_not_using_indexes=...).
A value of 0 disables the limit.

The first such warning opens a 60s timeframe; warnings up to 
the given limit are printed.
Any warnings beyond the given limit are suppressed.
A summary stating the number of suppressed warnings and the
total time spent in all queries we didn't print a warning for
will be logged at the end of the window. (This summary line
gives query-time and lock-time, just like a regular line in
the slow query log would; however, the values are cumulative
for all suppressed lines in that timeframe.)
The next 60s timeframe does not start immediately after a
summary is printed; instead, it starts when the next
qualifying query ("not using indexes") is slow-logged.