BUG#11745230
Post-push fix for the test in mysql-trunk.

The entry for the relay-log filename in relay-log.info file is on the
line and not on the first one as in mysql-5.5. This made the test case
fail because the relay-log.info was not correctly rewritten.

In addition, I've added the warnings due to the inclusion of 
master-slave.inc into the result file (after WL 4143, mysql-trunk 
throws these warnings when the password is stored in the master.info).