WL#4305 - storage-engine private data area per physical table
Implementing WL#4305 on the TABLE_SHARE and the handler class,
including the engines partitioning, InnoDB, Archive and Example.

Archive, InnoDB and Example is very similar, only differs in the
engine specific data.

Partitioning is similar, but extended with creating storage space for
each partition.

The main idea is to have a Handler_share class, which all engines can
inherit and add their own shared data to.

A pointer to the table's Handler_share is added to the TABLE_SHARE.
And given to the engine right after it has been created with get_new_handler.
If the TABLE_SHARE is given, then it must also call set_ha_share_storage
with a pointer to the TABLE_SHARE::ha_share. For partitions handlers
it is called by the partition handler pointing out an area in the Partition_share.

The partitioning alter code has been changed to not use a temporary
TABLE_SHARE, but instead use the original one to enforce only having
one TABLE_SHARE per table. The new/temporary partitions are given
a ha_share area within the partitioning handler, since they will only
last as long as the statement and the handler will be destroyed in the end
of the statement.
Due to this change some result files for crashing/failure during partition alter
was updated. Also test partition_myisam was updated since it did no longer
opened a temporary instance during alter.

Also cleaned up the crash/failure tests.