BUG#11745230
Making the test case work when relay-log.info data is stored in a 
table. 

The test case assumed that the relay-log.info data to be fixed was
stored in a file. However, it can be stored in a table as well. For
completeness, we fix this the test case to make it work in such cases 
as well.