Partial fix for Bug#11764622 57480: MEMORY LEAK WHEN HAVING 256+ TABLES
In row_create_prebuilt() try to estimate how many bytes will be allocated
from the mem_heap_t object and create it with the appropriate size so that
no further allocations are done by the code.

The only exception we make is that we do not allocate mysql_row_len bytes
if it is bigger than 256. This memory chunk is needed only in INSERTs and
since we are not sure what is the prebuilt going to be used for and since
mysql_row_len could be huge (as huge as e.g. 60KB) - we do not allocate
a space for it in advance.

This change removes at least this codepath:

#0 malloc (size=1384)
#1 mem_area_alloc
#2 mem_heap_create_block
#3 mem_heap_add_block
#4 mem_heap_alloc
#5 dtuple_create
#6 row_create_prebuilt
#7 ha_innobase::open
#8 handler::ha_open
#9 ha_partition::open
#10 handler::ha_open
#11 open_table_from_share
#12 open_table
#13 open_and_process_table
#14 open_tables
#15 open_and_lock_tables

causing at least 1 less malloc() during open table.

In exchange this codepath:

#0 malloc
#1 mem_area_alloc
#2 mem_heap_create_block
#3 mem_heap_create_func
#4 row_create_prebuilt
#5 ha_innobase::open
#6 handler::ha_open
#7 ha_partition::open
#8 handler::ha_open
#9 open_table_from_share
#10 open_table
#11 open_and_process_table
#12 open_tables
#13 open_and_lock_tables
#14 open_and_lock_tables
#15 execute_sqlcom_select

now allocates 3104 bytes instead of 632 (on 64 bit machine).

Reviewed by:	Inaam (rb:817)