Partial fix for Bug#11764622 57480: MEMORY LEAK WHEN HAVING 256+ TABLES
Make row_prebuilt_t::pcur and row_prebuilt_t::clust_pcur btr_pcur_t
instead of btr_pcur_t* in the definition of row_prebuilt_t.

It makes no sense to have pointers to objects that are always
(unconditionally) allocated separately from the main object. It is better
to allocate them along with the main object (row_prebuilt_t).

This is possible because we do not do any of the followings in the code:
1. Assign ::pcur or ::clust_pcur to some object allocated elsewhere
2. Free and reallocate ::pcur or ::clust_pcur

This change removes two malloc() calls per one open table. It removes the
following codepath (executed twice per open table):

#0 malloc (size=352)
#1 mem_area_alloc
#2 mem_heap_create_block
#3 mem_heap_create_func
#4 mem_alloc_func
#5 btr_pcur_create_for_mysql
#6 row_create_prebuilt
#7 ha_innobase::open
#8 handler::ha_open
#9 ha_partition::open
#10 handler::ha_open
#11 open_table_from_share
#12 open_table
#13 open_and_process_table
#14 open_tables
#15 open_and_lock_tables

and it changes this codepath to allocate 3552 instead of 3104 bytes:

#0 malloc
#1 mem_area_alloc
#2 mem_heap_create_block
#3 mem_heap_create_func
#4 row_create_prebuilt
#5 ha_innobase::open
#6 handler::ha_open
#7 ha_partition::open
#8 handler::ha_open
#9 open_table_from_share
#10 open_table
#11 open_and_process_table
#12 open_tables
#13 open_and_lock_tables
#14 open_and_lock_tables
#15 execute_sqlcom_select

Reviewed by:	Inaam (rb:817)