Partial fix for Bug#11764622 57480: MEMORY LEAK WHEN HAVING 256+ TABLES
In dict_load_foreigns() use onstack storage for the dtuple that is only
used inside that function. And do not use a separate storage for storing
a '\0'-terminated foreign key id, instead pass the len of the string to
dict_load_foreign() and change it to work with non-terminated string.

Introduce a new function dtuple_create_from_mem() that initializes a
dtuple from an already allocated memory chunk.

This change removes two malloc() calls per one open table (for a table
that does not have foreign keys!). It removes the following codepath
(executed twice per open table):

#0 malloc (size=376)
#1 mem_area_alloc
#2 mem_heap_create_block
#3 mem_heap_create_func
#4 dict_load_foreigns
#5 dict_load_table
#6 dict_table_open_on_name_low
#7 dict_table_open_on_name
#8 ha_innobase::open
#9 handler::ha_open
#10 ha_partition::open
#11 handler::ha_open
#12 open_table_from_share
#13 open_table
#14 open_and_process_table
#15 open_tables

and it causes the stack memory footprint of dict_load_foreigns() to grow
with 56 bytes (DTUPLE_EST_ALLOC(1)==72 minus two pointers, 8 bytes each).

Reviewed by:	Inaam (rb:817)