Partial fix for Bug#11764622 57480: MEMORY LEAK WHEN HAVING 256+ TABLES
In ha_innobase::open() we used to allocate two buffers: upd_buff and
key_val_buff. Both with the size of the entire row of the table. Also
in ha_innobase::records_in_range() we allocated yet another buffer
with the same size as key_val_buff.

With a table like 'c VARCHAR(64000), KEY (c(3072))' we used to allocate
about 128KB at table open time for those two buffers (and another 64KB
just for the duration of ha_innobase::records_in_range()).

upd_buff is only used in ha_innobase::update_row(), so we delay its
allocation until ha_innobase::update_row() is called. After all, that
method may never be called in the lifetime of the ha_innobase object.
Renamed to upd_buf on purpose - to trigger a compilation error if there
appears to be some other place where it is used and also for consistency
with other InnoDB code that uses 'buf' instead of 'buff'.

key_val_buff is only used in ha_innobase::index_read() and in
ha_innobase::records_in_range() to be passed as a working space to
row_sel_convert_mysql_key_to_innobase() (but the generated search tuple
may end up having a pointer inside that buf, so the lifetime of the buf
should be as long as the lifetime of the search tuple). That buffer does
not need to be the size of the whole row, only the part covered by indexes.
So we allocate the buffer directly in the ha_innobase object space. This
also makes the code cleaner.

This change removes the following malloc() during open table:

#0 malloc (size=128128)
#1 my_malloc
#2 my_multi_malloc
#3 ha_innobase::open
#4 handler::ha_open
#5 ha_partition::open
#6 handler::ha_open
#7 open_table_from_share
#8 open_table
#9 open_and_process_table
#10 open_tables
#11 open_and_lock_tables
#12 open_and_lock_tables
#13 execute_sqlcom_select
#14 mysql_execute_command
#15 mysql_parse

where the allocation size is roughly twice the row size.

In exchange if UPDATE is ever done on that table, then during the first
UPDATE an allocation of about the row size will be done (64k in this example).

Also the ha_innobase object will become about 6KB bigger due to allocating two
buffers of size REC_VERSION_56_MAX_INDEX_COL_LEN+2 (3072+2) along with it.

Reviewed by:	Inaam (rb:817)