Bug#12750920: EMBEDDED SERVER START/STOP
The main problem was pthread_key's was never deleted.

Solved by deleting them when stopping the embedded library.

Also fixed a couple of mem leaks found when running the test under
valgrind.

Some global sys_var was incompatible with restarts of the embedded
library. They where preallocated and released at the end of the
embedded library, but it then set its allocated flag to zero, resulting
in it would not be released on successive ends of the embedded
library. Solved by adding static buffers to use instead.

The test now restarts the embedded library > 1024 times.