Fix Bug#12316365 PERSISTENT STATISTICS FOR LARGE TABLE IS OFF BY SEVERAL MAGNITUDES
There was a bug in the sampling algorithm - the estimate for the number
of distinct records on the leaf level was estimated as:

number of distinct records on an upper level (LA) *
average number of distinct records on a leaf page

but it should be simply:

number of leaf pages *
average number of distinct records on a leaf page

This was unnoticed becasue most of the trees consist of 1, 2 or 3 levels
and for those we either did full scan or descended to the last but one level
(the one above the leaf) where the "number of distinct records" equals the
number of leaf pages (for a PK).

Remove the parameter total_recs_on_level from
dict_stats_analyze_index_for_n_prefix() since it is not needed anymore in
calculations - index->stat_n_leaf_pages must be used instead.