Bug#13418197: ASSERTION `(*TABLES)->REGINFO.LOCK_TYPE >=               TL_READ' FAILED | MYSQL_LOCK_TABLES
This assert was triggered if a view was created with
the same name as an existing temporary table.

The problem was that temporary tables are pre-opened when
CREATE VIEW starts. Later, when normal open_and_lock_tables()
is executed, CREATE VIEW specifies that only base tables
should be opened. However, since temporarary tables have
been pre-opened, this restriction has no effect. And if
a temporary table exist with the same name as the view,
the assert is later triggered during table locking.

This patch fixes the problem by not pre-opening temporary
tables when executing CREATE VIEW statements.

This bug was a regression caused by the patch for
Bug#11746602.

Test case added to view.test.