BUG#13427444: LIVE UPGRADE FAILS BETWEEN 5.5.18 -> 5.6.4
When doing a live upgrade from a 5.5 into a 5.6, the server would
refuse to startup because the new replication metadata tables
were not created.  In addition, the error message was mostly
unhelpful. In fact if the user is convinced that the server
should start, he can turn on the option slave-skip-start which
also bypasses the replication metadata checks on startup. Then
the user is able to fix the problem.
  
In this patch we extend the error and warning messages to make it
clear that the problem may be related to the live upgrade and
also to suggest using skip-slave-start, to bypass the startup
problem. The user can then resort to mysql_upgrade to fix not
only this but other problems in system tables that are due to the
live upgrade.