BUG#11766576 - 59715: UPDATE IGNORE, 1 ROW AFFECTED
BACKGROUND:
UPDATE IGNORE returns the wrong number of rows updated when there
are duplicate key conflicts in the multi table update.

FIX:
This bug has been fixed by putting a check on updated variable 
in mysql_multi_update::do_update function so that count is not
updated when duplicate key conflicts occur (and the corresponding 
rows are not updated).