Bug#12725206/Bug#61765: WRONG QUERY RESULTS ON SUBPARTITIONS                                                BASED ON USIGNED BIGINT
The pruning function get_next_subpartition_via_walking did always try to
store the value as signed, resulting in error when storing an unsigned
value, which then did not find any partition.

The solution was the same as bug#20257, but for subpartitions.
I.e. use the fields unsigned flag when storing the value.