BUG# 13417754 - ASSERT IN ROW_DROP_DATABASE_FOR_MYSQL DURING DROP SCHEMA
If an index is corrupt but we force it to be loaded into the dictionary cache
during DROP table then we don't move it to the non-LRU list but return NULL
(error) from dict_table_open_on_name_low().

This NULL return however causes the control to skip over the code that moves
the table from LRU to non-LRU in row_drop_table_for_mysql().

The fix is to move the table from LRU to non-LRU when we detect corruption
in dict_table_open_on_name_low().

Approved by Jimmy, rb://853