WL#4443 fixing query cache tests
Reverted back to storing queries in the query cache after
the tables was locked.

This makes transactional storage engines work with the query cache
since they can use the lock as signal to start a new transaktion.

This also means that queries that fails (like on access control) before
the tables are locked is not registered and is not included in
Qcache_not_cached status variable.