Bug#12589167 60901: 5.6.2 CLIENT TRUNCATES LINES AT              255 CHARS/MULTILINE QUERIES DONT WORK              ON WINDOWS
The synopsis of the bug does not outline the real
problem. The actual problem being the copy/paste,
i.e. whenever a multi-lined string is copied leaving
the trailing line-separator characters "CRLF" of the
last line, the first character of the last line
always gets altered to 'm', and that is the reason
behind this bug.

Fixed by changing the way lines are read from the
console buffer (initially done as part of WL#5331).
Now, the entire line will be read in a single go
instead of character-by-character reading.