Bug#13552968: Extra row with materialization on join + subquery in where +
The problem here occurs with an outer join between two "system" tables
(tables with exactly one row each), and with an "expensive" join
condition between the two tables. User-defined functions and materialized
subqueries are considered expensive in this context.

The fix for bug#52344 is related to this problem, as it moved subquery
materialization out of the optimizer phase and into the execution phase.
However, when marking the inner table of the outer join as "const",
it is not possible to later evaluate the join condition and subsequently
mark the fields from the inner tables as NULL. This is contrary to the
condition attached to the JOIN object: If there are only const tables
in the JOIN, do_select() will evaluate this condition and correctly
return an empty result if the condition evaluates to FALSE.

The cases for where the bug hit were quite interesting:
- With semijoin and materialization ON, the subquery is first marked
  as semi-join candidate, but because the outer query is OUTER JOIN,
  it is instead reverted to IN-TO-EXISTS. This is marked non-expensive
  and thus does not pose a problem.
- With semijoin OFF, materialization is selected. The condition is then
  marked as expensive and the problem occurs.
- With STRAIGHT JOIN applied to the outer query, materialization is
  always selected, and the problem occurs.

The chosen solution is to stop marking the inner table of an outer join
that has an expensive join condition as a const table.
Fixing the executor to handle this particular case does not seem viable.
 
mysql-test/include/subquery_mat.inc
  Added test case for bug#13552968.
  Modified test case for bug#52344 so that problem is visible here as well.

mysql-test/r/subquery_mat.result
mysql-test/r/subquery_mat_all.result
mysql-test/r/subquery_mat_none.result
  Added test case results for bug#13552968.

sql/sql_executor.cc
  Ensure that join_read_const_table() is not called with an outer-joined
  table with an expensive condition.

sql/sql_optimizer.cc
  For the inner table of an outer join that has an expensive condition,
  do not mark a table containing one row as const.