Bug#13578660 NDB REPLICATION : CONFLICT DETECTION NOT ENABLED ON ALL CONNECTED MYSQLDS
Conflict function setup was using a TABLE* object 
(MySQLD table representation) to determine the offsets
of the resolve column in the table record for NDB$MAX and
NDB$OLD, and the primary key column offsets for exceptions
table inserts.

Conflict function setup is part of binlog setup and occurs
as a result of :
 - Table creation
 - Table 'discovery' due to observed ndb_schema notifications
 - Table 'discovery' due to SHOW TABLES
 - Table 'discovery' at server startup

In some of these situations, the calling context has no MySQLD
Table* for the table, and so null is passed.  As the information
to setup conflict detection was not available, it was not set up!

Fix

Remove the use of a TABLE* when setting up conflict detection.
Parts :
 1) Store resolve column number in NDB_SHARE rather than a record
    offset
 2) Store exception's table pk attrids in NDB_SHARE rather than
    record offsets
 3) Modify conflict function definition code to use resolve column
    number and NdbRecord to determine the correct offset.
 4) Modify exceptions table insert code to use pk attrids and 
    NdbRecord to determine the correct offsets.
 5) Avoid using TABLE* members for error messages etc
 6) Remove TABLE* parameter from calls
 7) Remove check for TABLE*==NULL

Additional items

Cleanup/improve reporting of conflict function setup.
Use sufficient bits to encode resolve column attrid.
Add testcase to verify conflict table setup distribution.