Fix late definition problem
ndb_thd.cc #defines MYSQL_SERVER just before it #includes sql_class.h.
Type definitions with sql_class.h change size and offsets depending on
the definition of MYSQL_SERVER (Particularly the NET type embedded in
THD).
Unfortunately sql_class.h #includes other headers, which also depend on
MYSQL_SERVER and which have already been included, without it defined,
prior to this point.
This results in an invalid definition of THD for code after the #include
of sql_class.h, which can result in data corruption during execution.

The fix is to move the MYSQL_SERVER #define up to the earliest point.

Testing is difficult as it is not easy to hit the problem (execution
of thd_print_warning_list()) without error injection.  However, a separate
feature to follow adds testing of this path.