Bug#13500371 63704: CONVERSION OF '1.' TO A NUMBER GIVES ERROR 1265 (WARN_DATA_TRUNCATED)
Problem: Converting strings ending with a decimal point (such as '1.')
to floating point numbers causes a truncation warning to be emitted.

Field_float::store() and Field_double::store() calls my_strtod_int()
to convert strings to floating point numbers. Function my_strtod_int()
doesn't consume the decimal point if it is the last character of the
string, and this leads to Field_float::store() and
Field_double::store() reporting unconsumed character as truncated
data.

Fix: Let my_strtod_int() consume the decimal point also if it is the
last character of the string. This behavior is consistent with that of
scanf().