Bug #13604034 SET GLOBAL INNODB_UNDO_LOGS=0 SUCCEEDS BUT LEADS TO AN ASSERT MAX_UNDO_LOGS > 0
The min value was not checked, while setting the value of innodb_undo_logs config
value.  This was fixed.  Also, the max value and the default value is not fixed in
the case of innodb_undo_logs.  So after starting innodb, the max and default value
of the innodb_undo_logs config will be reset to the proper value based on the 
srv_available_undo_logs variable.  

This system variable framework does not support doing the validation for default
value (because the boundaries min and max is not fixed).  Because of this limitation
of the framework, we needed to adjust the max and default value of the config object.