WL#4443 - Additional boost pack for heavily partitioned tables.
This is the low hanging fruits of performance patches that was discovered
after WL4443 was completed.

Implemented bitmap_get_next_set and added it to the unittest.
This makes scanning through empty partitions much faster.
Very good for primary key/partition key and explicit prune
operations.

Added m_started_partitions, which tracks which partitions that
are started for the current statement
(either in external_lock(!=F_UNLCK) or in start_stmt).
This prunes ha_partition::reset() to only call partitions
that where 'started'. Good for prepare statement, which
calls ha_reset() after prepare, without ever starting anything.

Changed m_num_locks to be number of locks per partition,
instead of locks for all partitions. This is now used in
ha_partition::store_lock() to return the number of locks to be
used after pruning, instead of the old way, which was to return
the sum of locks needed for all partitions. This uses a bit less
resources and makes the get_lock_data() faster
(part of mysql_lock_tables).

Skipping second call to prune_partitions (in the optimize step) if it
only has constant items and no subqueries, since it will not prune
anything more than it did in the prepare stage.
(prune_partitions takes several percent of the time for non complex
statements, like pk lookups.) One has to do the pruning in the prepare
stage, since it can prune external_lock/start_stmt.
So even if it costs more in prepare than it does in optimize,
the gain of pruning external_lock/start_stmt  is so much bigger.

Disabled PSI_CALL(open_table)() call in ha_open and removed
rebind_psi/unbind_psi for partitions. Since this is done
when a table is opened, taken from the table cache or closed
and cannot be pruned, since pruning is done after open.
This avoids using too much resources and increase performance
for heavily partitioned tables, but will not remove the tracing of
table io for each partition.

Refactored the init/free of the internal bitmaps over which
partitions that have been locked, started or called bulk_insert.