BUG#13599013 - MAX/MIN + SUBQUERY IN WHERE CLAUSE MATCHING NO                ROWS + INDEX DOES NOT RETURN NULL
An aggregated expression of type MIN or MAX shall return NULL
if the query is implicitly grouped and there is no HAVING 
clause that evaluates to FALSE.  In this bug we had a query 
of the form

  SELECT MIN(indexed_column) FROM t1 WHERE (4) IN (SELECT 1)

No row can satisfy the WHERE condition but there is no HAVING
clause, so the query should have returned NULL (but did
return an empty set).

The problem was that when a query only selects aggregate
functions that are optimized away, JOIN::exec() goes into an
early-exit branch. This branch did not honor the requirement
stated by the definition above. The fix is to call
return_zero_rows() instead of send_eof() in this branch. If
the query selects aggregates and does not have a HAVING
clause that evaluates to false, return_zero_rows() returns
a NULL-based row before calling send_eof().