Fix for bug#13528826  TEST_IF_CHEAPER_ORDERING(): CALCULATES INCORRECT 'SELECT_LIMIT
An incorrect 'adjusted select_limit' was returned from test_if_cheaper_ordering()
if no excplicit limit-clause was specified in the query.

This fix detect when a limit-clause was not specified, and returns the
full 'table_records' when no limit-clause was used.

Furthermore, this fix caused the testcase for bug #12838420 
to change beyond what was required as a testcase for that bug.
      
Fixed that by forcing correct query plan for testcase by
introducing 'USE INDEX' and 'STRAIGHT_JOIN'.